services/video-crawler/requirements.txt
alembic==1.4.2
astroid==2.4.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
dill==0.3.3
idna==2.9
importlib-metadata==1.6.0
isort==5.6.4
kafka-python==2.0.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyfunctional==1.4.2
pylint==2.6.0
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
tabulate==0.8.7
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0
